Clippy

+++ crates/common/types/blobs_bundle.rs
@@ -378,8 +378,7 @@ mod tests {
     fn transaction_with_too_many_blobs_should_fail() {
         let blob = blobs_bundle::blob_from_bytes("Im a Blob".as_bytes().into())
             .expect("Failed to create blob");
-        let blobs =
-            std::iter::repeat_n(blob, MAX_BLOB_COUNT_ELECTRA as usize + 1).collect::<Vec<_>>();
+        let blobs = std::iter::repeat_n(blob, MAX_BLOB_COUNT_ELECTRA + 1).collect::<Vec<_>>();
 
         let blobs_bundle =
             BlobsBundle::create_from_blobs(&blobs).expect("Failed to create blobs bundle");
